requirements.txt
uvicorn==0.31.0
fastapi==0.115.0
python-dotenv==1.0.1
pymongo==4.10.1
python-dateutil==2.8.2
python-multipart
notion-client
